Fix iterator invalidation on erase
Before the fix if we erase a node, which has both children we would get
its inorder successor and basically swap the 'data_' value between the
nodes. This will lead to iterator invalidation, since we remove the
other node. After the fix we swap the nodes, not just the 'data_' value.
Also add tests for this case.